+++ Trarizon.Library/History.md
@@ -9,6 +9,7 @@
     - Optimize `AllocOptSet<>` with `StackAllocBitArray`
     - Optimize `IEnumerable<>.TrySingle()`, `TryFirst()`
     - Remove unused method in `IKeyedEqualityComparer`(internal)
+    - Optimize internal structure of `AllocOptSet<>`
 
 ## v0.2.1